Updating framework version and refactoring package names.
