+++ components/org.wso2.carbon.identity.application.authenticator.iwa.ntlm/src/main/java/org/wso2/carbon/identity/application/authenticator/ntlm/IWAConstants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.application.authenticator.iwa;
+package org.wso2.carbon.identity.application.authenticator.ntlm;
 
 public class IWAConstants {
