+++ components/org.wso2.carbon.identity.application.authenticator.iwa.ntlm/src/main/java/org/wso2/carbon/identity/application/authenticator/ntlm/IWAServiceDataHolder.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.identity.application.authenticator.iwa;
+package org.wso2.carbon.identity.application.authenticator.ntlm;
 
 import waffle.servlet.spi.SecurityFilterProviderCollection;
 import waffle.windows.auth.IWindowsAuthProvider;
