+++ pom.xml
@@ -24,9 +24,9 @@
         <version>1</version>
     </parent>
 
-    <groupId>org.wso2.carbon.identity</groupId>
-    <artifactId>identity-application-auth-iwa</artifactId>
-    <version>5.1.2</version>
+    <groupId>org.wso2.carbon.identity.application.auth.iwa.ntlm</groupId>
+    <artifactId>identity-application-auth-iwa-ntlm</artifactId>
+    <version>5.2.0-SNAPSHOT</version>
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
     <name>WSO2 Carbon - IWA Identity Application Authenticator Module</name>
@@ -36,17 +36,16 @@
     <url>http://wso2.org</url>
 
     <scm>
-        <url>https://github.com/wso2-extensions/identity-local-auth-iwa.git</url>
-        <developerConnection>scm:git:https://github.com/wso2-extensions/identity-local-auth-iwa.git
+        <url>https://github.com/wso2-extensions/identity-local-auth-iwa-ntlm.git</url>
+        <developerConnection>scm:git:https://github.com/wso2-extensions/identity-local-auth-iwa-ntlm.git
         </developerConnection>
-        <connection>scm:git:https://github.com/wso2-extensions/identity-local-auth-iwa.git</connection>
+        <connection>scm:git:https://github.com/wso2-extensions/identity-local-auth-iwa-ntlm.git</connection>
         <tag>v5.1.2</tag>
     </scm>
 
     <modules>
-        <module>components/org.wso2.carbon.identity.application.authenticator.iwa</module>
-        <!--<module>components/org.wso2.carbon.identity.authenticator.iwa.stub</module>-->
-        <module>features/org.wso2.carbon.identity.application.authenticator.iwa.server.feature</module>
+        <module>components/org.wso2.carbon.identity.application.authenticator.iwa.ntlm</module>
+        <module>features/org.wso2.carbon.identity.application.authenticator.iwa.ntlm.server.feature</module>
     </modules>
 
     <repositories>
@@ -139,12 +138,12 @@
 
             <!--Carbon Identity Framework Dependencies-->
             <dependency>
-                <groupId>org.wso2.carbon.identity</groupId>
+                <groupId>org.wso2.carbon.identity.framework</groupId>
                 <artifactId>org.wso2.carbon.identity.application.authentication.framework</artifactId>
                 <version>${carbon.identity.framework.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.carbon.identity</groupId>
+                <groupId>org.wso2.carbon.identity.framework</groupId>
                 <artifactId>org.wso2.carbon.identity.application.common</artifactId>
                 <version>${carbon.identity.framework.version}</version>
             </dependency>
@@ -247,7 +246,7 @@
 
     <properties>
         <!--Carbon Identity Framework Version-->
-        <carbon.identity.framework.version>5.2.1</carbon.identity.framework.version>
+        <carbon.identity.framework.version>5.6.14</carbon.identity.framework.version>
         <carbon.identity.framework.import.version.range>[5.0.0, 6.0.0)</carbon.identity.framework.import.version.range>
 
         <!-- Carbon Identity Local Authenticator IWA version -->
